Modify the CurrentPriceEntity
* Modified the currentPriceEntity
* Added variables and variable type from the float to double
* Modified the CurrentPriceModel by entity's change
* Added a DynamicHeightFlowLayout

+++ SiseWebSocketSample/Model/CurrentPriceModel.swift
@@ -9,9 +9,9 @@ import Foundation
 
 struct CurrentPriceModel {
     var stockName: String
-    var currentPrice: Float
-    var percentChange: Float
-    var prevPriceRate: Float
+    var currentPrice: Double
+    var percentChange: Double
+    var prevPriceRate: Double
     var isUp: Bool
     var symbol: String
 }
